refactor(rank): separated files and removed pointers
